Moved float comparison to vector utils
