+++ pkg/vectors/utils.go
@@ -1,17 +1,19 @@
-package numeric
+package vectors
 
-import "math"
+import (
+	"math"
+)
 
 // Float number accuracy at the project boundaries.
-const AbsoluteTolerance = 1e-12
+const absoluteTolerance = 1e-12
 
 // Main float number comparison method. It must be used in any
 // computation or condition to avoid float value tricks and traps.
-func IsClose(a, b float64) bool {
-	return math.Abs(a-b) <= AbsoluteTolerance
+func isClose(a, b float64) bool {
+	return math.Abs(a-b) <= absoluteTolerance
 }
 
 // A shortcut for float comparison with zero.
-func IsNull(a float64) bool {
-	return IsClose(a, 0)
+func isNull(a float64) bool {
+	return isClose(a, 0)
 }
